Organized components in src directory; deleted empty/test files
