+++ src/components/contact/desktop/ContactForm.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { useForm } from "react-hook-form";
 import {ButtonBase, Grid, makeStyles, Snackbar, styled, TextField, Typography} from '@material-ui/core'
-import { theme } from "../theme";
+import { theme } from "../../../theme";
 import {AccountCircleTwoTone, EmailTwoTone, RateReviewTwoTone} from '@material-ui/icons';
 
 // Airtable integration
