--- src/components/icons/ArrowBracketsIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-
-export const ArrowBracketsIcon: React.FC = () => {
-
-  return (<div/>);
-}
