--- src/components/icons/BackgroundPersonIcon.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-
-export const BackgroundPersonIcon: React.FC = () => {
-
-  return (<div/>);
-}
