--- src/components/icons/CheckDotIcon.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-
-export const CheckDotIcon: React.FC = () => {
-
-  return (<div/>);
-}
