--- src/components/icons/ChevronDownIcon.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-
-export const ChevronDownIcon: React.FC = () => {
-
-  return (<div/>);
-}
