--- src/components/icons/CircleWithSliverIcon.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-
-export const CircleWithSliverIcon: React.FC = () => {
-
-  return (<div/>);
-}
