--- src/components/icons/ConcentricDotIcon.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-
-export const ConcentricDotIcon: React.FC = () => {
-
-  return (<div/>);
-}
