--- src/components/icons/DOrgLogo.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-
-export const DOrgLogo: React.FC = () => {
-
-  return (<div/>);
-}
