--- src/components/icons/DashIcon.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-
-export const DashIcon: React.FC = () => {
-
-  return (<div/>);
-}
