--- src/components/icons/GearIcon.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-
-export const GearIcon: React.FC = () => {
-
-  return (<div/>);
-}
