--- src/components/icons/GearIcon2.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-
-export const GearIcon2: React.FC = () => {
-
-  return (<div/>);
-}
