--- src/components/icons/GitHubIcon.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-
-export const GitHubIcon: React.FC = () => {
-
-  return (<div/>);
-}
