--- src/components/icons/JewelIcon.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-
-export const JewelIcon: React.FC = () => {
-
-  return (<div/>);
-}
