--- src/components/icons/KeybaseIcon.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-
-export const KeyBaseIcon: React.FC = () => {
-
-  return (<div/>);
-}
