--- src/components/icons/LeftHandShakeIcon.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-
-export const LeftHandShakeIcon: React.FC = () => {
-
-  return (<div/>);
-}
