--- src/components/icons/MailIcon.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-
-export const MailIcon: React.FC = () => {
-
-  return (<div/>);
-}
