--- src/components/icons/PenTipIcon.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-
-export const PenTipIcon: React.FC = () => {
-
-  return (<div/>);
-}
