--- src/components/icons/QuarterCircleIcon.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-
-export const QuarterCircleIcon: React.FC = () => {
-
-  return (<div/>);
-}
