--- src/components/icons/RightArrowIcon.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-
-export const RightArrowIcon: React.FC = () => {
-
-  return (<div/>);
-}
