--- src/components/icons/SimpleDotIcon.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-
-export const SimpleDotIcon: React.FC = () => {
-
-  return (<div/>);
-}
