--- src/components/icons/SmileyFaceIcon.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-
-export const SmileyFaceIcon: React.FC = () => {
-
-  return (<div/>);
-}
