--- src/components/icons/SquareLoopIcon.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-
-export const SquareLoopIcon: React.FC = () => {
-
-  return (<div/>);
-}
