--- src/components/icons/TinyDotIcon.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-
-export const TinyDotIcon: React.FC = () => {
-
-  return (<div/>);
-}
