--- src/components/icons/TwitterIcon.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-
-export const TwitterIcon: React.FC = () => {
-
-  return (<div/>);
-}
