--- src/components/icons/VerticalDashIcon.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-
-export const VerticalDashIcon: React.FC = () => {
-
-  return (<div/>);
-}
