+++ src/pages/Home.tsx
@@ -2,7 +2,7 @@
 import { Box, styled } from "@material-ui/core";
 import React from "react";
 import ReactGA from "react-ga";
-import {Communitycomp} from "../components/CommunityComp"
+import {Communitycomp} from "../components/home/desktop/CommunityComp"
 const Root = styled(Box)({
   margin: 'auto'
 });
